chore(plugins/plugin-client-common): refactor all markdown components into individual files
