+++ plugins/plugin-client-common/src/components/Content/Markdown/components/list.tsx
@@ -28,7 +28,7 @@ import {
   liveStatusChannel
 } from '../../ProgressStepper'
 
-export function listItem(props: LiProps) {
+export function li(props: LiProps) {
   if (isProgressStepCompatible(props)) {
     return (
       <ProgressStep
