+++ plugins/plugin-client-common/src/components/Content/Markdown/components/table.tsx
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2021 The Kubernetes Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react'
+
+export function table(props: React.TableHTMLAttributes<HTMLTableElement>) {
+  return <table className={props.className + ' kui--table-like kui--structured-list'}>{props.children}</table>
+}
+
+export function thead(props: React.HTMLAttributes<HTMLTableSectionElement>) {
+  return <thead className={props.className + ' kui--structured-list-thead'}>{props.children}</thead>
+}
+
+export function tbody(props: React.HTMLAttributes<HTMLTableSectionElement>) {
+  return <tbody className={props.className + ' kui--structured-list-tbody'}>{props.children}</tbody>
+}
